+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/SchemaUtils.kt
@@ -614,17 +614,6 @@ object SchemaUtils {
     }
 
     /**
-     * Log Exposed table mappings <-> real database mapping problems and returns DDL Statements to fix them, including
-     * DROP/DELETE statements (unlike [checkMappingConsistence])
-     */
-    private fun mappingConsistenceRequiredStatements(vararg tables: Table, withLogs: Boolean = true): List<String> {
-        return checkMissingIndices(tables = tables, withLogs).flatMap { it.createStatement() } +
-            checkUnmappedIndices(tables = tables, withLogs).flatMap { it.dropStatement() } +
-            checkExcessiveForeignKeyConstraints(tables = tables, withLogs).flatMap { it.dropStatement() } +
-            checkExcessiveIndices(tables = tables, withLogs).flatMap { it.dropStatement() }
-    }
-
-    /**
      * Checks all [tables] for any that have more than one defined index and logs the findings. If found, this function
      * also logs the SQL statements that can be used to drop these indices.
      *
@@ -777,158 +766,6 @@ object SchemaUtils {
             indexes.subtract(nameDiffers).log("Indices exist in database and not mapped in code on class '$name':")
         }
         return toCreate.toList()
-    }
-
-    /**
-     * Checks all [tables] for any that have indices that are missing in the database but are defined in the code. If
-     * found, this function also logs the SQL statements that can be used to create these indices.
-     *
-     * @return List of indices that are missing and can be created.
-     */
-    private fun checkMissingIndices(vararg tables: Table, withLogs: Boolean): List<Index> {
-        fun Collection<Index>.log(mainMessage: String) {
-            if (withLogs && isNotEmpty()) {
-                exposedLogger.warn(joinToString(prefix = "$mainMessage\n\t\t", separator = "\n\t\t"))
-            }
-        }
-
-        val fKeyConstraints = currentDialect.columnConstraints(*tables).keys
-        val existingIndices = currentDialect.existingIndices(*tables)
-
-        fun List<Index>.filterForeignKeys() = if (currentDialect is MysqlDialect) {
-            filterNot { it.table to LinkedHashSet(it.columns) in fKeyConstraints }
-        } else {
-            this
-        }
-
-        // SQLite: indices whose names start with "sqlite_" are meant for internal use
-        fun List<Index>.filterInternalIndices() = if (currentDialect is SQLiteDialect) {
-            filter { !it.indexName.startsWith("sqlite_") }
-        } else {
-            this
-        }
-
-        fun Table.existingIndices() = existingIndices[this].orEmpty().filterForeignKeys().filterInternalIndices()
-
-        fun Table.mappedIndices() = this.indices.filterForeignKeys().filterInternalIndices()
-
-        val missingIndices = HashSet<Index>()
-        val nameDiffers = HashSet<Index>()
-
-        tables.forEach { table ->
-            val existingTableIndices = table.existingIndices()
-            val mappedIndices = table.mappedIndices()
-
-            for (index in existingTableIndices) {
-                val mappedIndex = mappedIndices.firstOrNull { it.onlyNameDiffer(index) } ?: continue
-                if (withLogs) {
-                    exposedLogger.info(
-                        "Index on table '${table.tableName}' differs only in name: in db ${index.indexName} -> in mapping ${mappedIndex.indexName}"
-                    )
-                }
-                nameDiffers.add(index)
-                nameDiffers.add(mappedIndex)
-            }
-
-            missingIndices.addAll(mappedIndices.subtract(existingTableIndices))
-        }
-
-        val toCreate = missingIndices.subtract(nameDiffers)
-        toCreate.log("Indices missed from database (will be created):")
-        return toCreate.toList()
-    }
-
-    /**
-     * Checks all [tables] for any that have indices that exist in the database but are not mapped in the code. If
-     * found, this function also logs the SQL statements that can be used to drop these indices.
-     *
-     * @return List of indices that are unmapped and can be dropped.
-     */
-    private fun checkUnmappedIndices(vararg tables: Table, withLogs: Boolean): List<Index> {
-        fun Collection<Index>.log(mainMessage: String) {
-            if (withLogs && isNotEmpty()) {
-                exposedLogger.warn(joinToString(prefix = "$mainMessage\n\t\t", separator = "\n\t\t"))
-            }
-        }
-
-        val foreignKeyConstraints = currentDialect.columnConstraints(*tables).keys
-        val existingIndices = currentDialect.existingIndices(*tables)
-
-        fun List<Index>.filterForeignKeys() = if (currentDialect is MysqlDialect) {
-            filterNot { it.table to LinkedHashSet(it.columns) in foreignKeyConstraints }
-        } else {
-            this
-        }
-
-        // SQLite: indices whose names start with "sqlite_" are meant for internal use
-        fun List<Index>.filterInternalIndices() = if (currentDialect is SQLiteDialect) {
-            filter { !it.indexName.startsWith("sqlite_") }
-        } else {
-            this
-        }
-
-        fun Table.existingIndices() = existingIndices[this].orEmpty().filterForeignKeys().filterInternalIndices()
-
-        fun Table.mappedIndices() = this.indices.filterForeignKeys().filterInternalIndices()
-
-        val unmappedIndices = HashMap<String, MutableSet<Index>>()
-        val nameDiffers = HashSet<Index>()
-
-        tables.forEach { table ->
-            val existingTableIndices = table.existingIndices()
-            val mappedIndices = table.mappedIndices()
-
-            for (index in existingTableIndices) {
-                val mappedIndex = mappedIndices.firstOrNull { it.onlyNameDiffer(index) } ?: continue
-                nameDiffers.add(index)
-                nameDiffers.add(mappedIndex)
-            }
-
-            unmappedIndices.getOrPut(table.nameInDatabaseCase()) {
-                hashSetOf()
-            }.addAll(existingTableIndices.subtract(mappedIndices))
-        }
-
-        val toDrop = mutableSetOf<Index>()
-        unmappedIndices.forEach { (name, indices) ->
-            toDrop.addAll(
-                indices.subtract(nameDiffers).also {
-                    it.log("Indices exist in database and not mapped in code on class '$name':")
-                }
-            )
-        }
-        return toDrop.toList()
-    }
-
-    /**
-     * Returns the SQL statements that need to be executed to make the existing database schema compatible with
-     * the table objects defined using Exposed. Unlike [statementsRequiredToActualizeScheme], DROP/DELETE statements are
-     * included.
-     *
-     * **Note:** Some dialects, like SQLite, do not support `ALTER TABLE ADD COLUMN` syntax completely,
-     * which restricts the behavior when adding some missing columns. Please check the documentation.
-     *
-     * By default, a description for each intermediate step, as well as its execution time, is logged at the INFO level.
-     * This can be disabled by setting [withLogs] to `false`.
-     */
-    fun statementsRequiredForDatabaseMigration(vararg tables: Table, withLogs: Boolean = true): List<String> {
-        val (tablesToCreate, tablesToAlter) = tables.partition { !it.exists() }
-        val createStatements = logTimeSpent("Preparing create tables statements", withLogs) {
-            createStatements(tables = tablesToCreate.toTypedArray())
-        }
-        val alterStatements = logTimeSpent("Preparing alter table statements", withLogs) {
-            addMissingColumnsStatements(tables = tablesToAlter.toTypedArray(), withLogs)
-        }
-
-        val modifyTablesStatements = logTimeSpent("Checking mapping consistence", withLogs) {
-            mappingConsistenceRequiredStatements(
-                tables = tables,
-                withLogs
-            ).filter { it !in (createStatements + alterStatements) }
-        }
-
-        val allStatements = createStatements + alterStatements + modifyTablesStatements
-        return allStatements
     }
 
     /**
