+++ exposed-java-time/src/test/kotlin/org/jetbrains/exposed/JavaTimeTests.kt
@@ -601,7 +601,7 @@ class JavaTimeTests : DatabaseTestsBase() {
             val date: Column<LocalDate> = date("date").index().defaultExpression(CurrentDate)
         }
         withTables(testTable) {
-            val statements = SchemaUtils.statementsRequiredForDatabaseMigration(testTable)
+            val statements = MigrationUtils.statementsRequiredForDatabaseMigration(testTable)
             assertTrue(statements.isEmpty())
         }
     }
