+++ exposed-kotlin-datetime/build.gradle.kts
@@ -22,6 +22,7 @@ dependencies {
     testImplementation(project(":exposed-dao"))
     testImplementation(project(":exposed-tests"))
     testImplementation(project(":exposed-json"))
+    testImplementation(project(":exposed-migration"))
     testImplementation(libs.junit)
     testImplementation(kotlin("test-junit"))
 }
